CMS-11508 Use "npm ci" instead of "npm install"
The devDependencies are not used by the production build and can be
removed. The package-lock.json file now only contains the actual
4 dependencies we need in the production build, which can be installed
super quick using "npm ci".

The version "1.06.1" of the textselection plugin caused problems
with an out-of-sync package-lock.json file since NPM expects a proper
semantic version. Using "1.6.1" instead fixed that.