ENT-5147 Bump parent hippo-cms7-project to 15.0.1
Since the introduction of the monorepo we don't have a separate
hippo-cms7-project POM anymore, which makes this <parent> reference
a bit strange.

The same CKEditor artifact is used in v13, v14, v15 and SAAS, which
is not an issue since it only contains javascript/css/html files that
are used by CKEditor, so it is mainly the build itself that requires
a <parent> POM (describing which version of NodeJS to use, a couple of
Maven plugins, etc.).

The hippo-cms7-project POM for v13 still references Node 12 and NPM 6,
which I did not want to use anymore, so instead the <parent> now
references the hippo-cms-project POM from v15.